fixed bug where labels were tokenized as symbols
Previously, keywords were being matched before symbols, because
otherwise all keywords would be tokenized as symbols. This
created a bug where symbols that start with a keyword were
getting tokenized as a keyword and then a symbol. This version
fixes that by adding a second phase to symbol matching that
checks for keywords.